requirements.txt
bluepy==1.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
flask==1.1.1
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-datastore==1.10.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
idna==2.9
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
